move protectionutil pkg to controller/util
