+++ pkg/controller/util/protectionutil/utils.go
