+++ pkg/controller/util/protectionutil/wrappers.go
